refactor: Adjust contents to center of screen
