+++ younghyun/src/components/responsive/centerAligner.jsx
@@ -3,7 +3,8 @@ import styled from "styled-components";
 const Container = styled.div`
     width: 100vw;
     max-width: ${(props) => props.maxWidth};
-    height: 100%;
+    height: 100vh;
+    max-height: 900px;
 `;
 
 function CenterAligner({ children, maxWidth }) {
