+++ younghyun/src/components/responsive/globalStyles.js
@@ -2,9 +2,10 @@ import { createGlobalStyle } from "styled-components";
 
 const GlobalStyle = createGlobalStyle`
     body {
-        display: flex;
         height: 100vh;
+        display: flex;
         justify-content:center;
+        align-items:center;
         background-color: #f2f2f2;
         color: #070707;
         margin: 0;
