+++ younghyun/src/components/responsive/mobileStyles.js
@@ -2,6 +2,8 @@ import { createGlobalStyle } from "styled-components";
 
 const MobileStyle = createGlobalStyle`
     body {
+        display: flex;
+        justify-content:center;
         background-color: #f2f2f2;
         color: #070707;
         margin: 0;
